Use embedded synclink tasks in the worker channel
